changing SocialNetworks@provider_id to decimal to allow storing google user's id

+++ database/migrations/2020_03_22_222107_create_social_networks_table.php
@@ -13,7 +13,7 @@ class CreateSocialNetworksTable extends Migration
             $table->id();
             $table->foreignId('users_id');
             $table->enum('provider_name', ['facebook', 'twitter', 'google']); 
-            $table->unsignedBigInteger('provider_id');
+            $table->decimal('provider_id', 65, 0);
             $table->string('username')->nullable();
             $table->string('profile_url')->nullable();
             $table->string('picture_url');
